Extract validation summary into hook
